:arrow_up: Updated dependencies, switched MDX plugin, and fixed #25
We were lagging behind a bit on the .dev site in comparison to the .org
site so I've upgraded all dependencies.
That includes the switch to the gatsby-plugin-mdx and the changes that
requires.

We're sticking with react-intl 2.x for now, because 3.x has a bunch of
breaking changes. There's an issue open for the migration on the .org
site so we'll do that first and then follow here.

I've also added previous/next navigation which closes #25.